Crate bip_util

Utilities used by the Bittorrent Infrastructure Project.

Modules

bt
Bittorrent specific types.

contiguous
Arrays of buffers as a contiguous buffer.

convert
Converting between data.

error
Common error types.

net
Networking primitives and helpers.

send
Generic sender utilities.

sha
Hash primitives and helpers.

test
Testing fixtures for dependant crates. TODO: Some non test functions in other crates use this, mark that as cfg test when we migrate away from these functions in non test functions.

trans
Generating transaction ids.

Functions

fisher_shuffle
Applies a Fisher-Yates shuffle on the given list.