Struct bip_util::contiguous::ContiguousBuffers

pub struct ContiguousBuffers<T> { /* fields omitted */ }

Struct for providing a ContiguousBuffer abstraction over many contiguous buffers.

Methods

impl<T> ContiguousBuffers<T>

fn new() -> ContiguousBuffers<T>

Create a new empty ContiguousBuffers struct.

fn with_buffer(buffer: T) -> ContiguousBuffers<T>

Create a new ContiguousBuffers struct with an initial element.

fn pack(&mut self, buffer: ContiguousBuffers<T>)

Pack a value T at the end of the contiguous buffers.

fn unpack<F>(self, accept: F) where
F: FnMut(ContiguousBuffers<T>),

Unpack all values T and pass them to the given closure.