Trait bip_util::contiguous::ContiguousBuffer

pub trait ContiguousBuffer<T> {
    fn capacity(&self) -> usize;
    fn length(&self) -> usize;
    fn clear(&mut self);
    fn write(&mut self, data: &[T]);
    fn read<F>(&self, receive: F)
    where
        F: FnMut(&[T]);
}

Trait for metadata, reading, and writing to a contiguous buffer that doesn't re allocate.

Required Methods

fn capacity(&self) -> usize

Total capacity of the underlying buffer.

fn length(&self) -> usize

Total length, or used capacity, of the underlying buffer.

fn clear(&mut self)

Clear the used capacity of the underlying buffer without re allocating.

fn write(&mut self, data: &[T])

Write to the unused capacity of the buffer.

Panic will occur if more data is given than capacity is available for.

fn read<F>(&self, receive: F) where
F: FnMut(&[T]),

Read the used capacity of the buffer.

Implementations will never pass in an empty slice.