Module bip_util::send

Generic sender utilities.

Structs

SplitSender
SplitSender allows dividing the capacity of a single channel into multiple channels.

SplitSenderAck
SplitSenderAck allows a client to ack messages received from a SplitSender.

Traits

TrySender
Trait for generic sender implementations.

Functions

split_sender
Create two SplitSenders over a single Sender with corresponding capacities.