Struct bip_util::send::SplitSender

pub struct SplitSender<S> { /* fields omitted */ }

SplitSender allows dividing the capacity of a single channel into multiple channels.

Methods

impl<S> SplitSender<S>

fn new(send: S, capacity: usize) -> SplitSender<S>

Create a new SplitSender.

fn sender_ack(&self) -> SplitSenderAck

Create a new SplitSenderAck that can be used to ack sent messages.