Trait bip_util::send::TrySender

pub trait TrySender<T: Send>: Send {
    fn try_send(&self, data: T) -> Option<T>;
}

Trait for generic sender implementations.

Required Methods

fn try_send(&self, data: T) -> Option<T>

Send data through the concrete channel.

If the channel is full, return the data back to the caller; if the channel has hung up, the channel should NOT return the data back to the caller but SHOULD panic as hang ups are considered program logic errors.