Module bip_util::sha

Hash primitives and helpers.

Structs

Bits
Iterator over some bits.

ShaHash
SHA-1 hash wrapper type for performing operations on the hash.

ShaHashBuilder
Building ShaHash objects by adding byte slices to the hash.

XorBits
Iterator over the bits of a xor operation.

Enums

BitRep
Representation of a bit.

XorRep
Representation of a bit after a xor operation.

Constants

SHA_HASH_LEN
Length of a SHA-1 hash.