Module bip_util::test

Testing fixtures for dependant crates. TODO: Some non test functions in other crates use this, mark that as cfg test when we migrate away from these functions in non test functions.

Functions

dummy_block_node_ids
Generates a block of unique dummy node ids as Vec<NodeId>

dummy_block_socket_addrs
Generates a block of unique ipv4 addresses as Vec<SocketAddr>

dummy_ipv4_addr
Generates a dummy Ipv4 address as an IpAddr.

dummy_ipv6_addr
Generates a dummy ipv6 address as an IpAddr.

dummy_node_id
Generates a dummy node id as a NodeId.

dummy_socket_addr_v4
Generates a dummy socket address v4 as a SocketAddr.

travel_into_future
Allows us to time travel into the future.

travel_into_past
Allows us to time travel into the past.