Struct bip_util::trans::LocallyShuffledIds

pub struct LocallyShuffledIds<T> { /* fields omitted */ }

Provides transaction ids that are locally shuffled.

For all intents and purposes, this class is used to generate transaction ids that need to be random but need to be unique over a reasonable span of time.

We do this by composing the sequential ids and taking x number of ids into a buffer. We then randomly shuffle that buffer and hand out ids in order. When the buffer is exhausted we repeat. This allows us to uphold the uniqueness property for any large transaction type (such as u64) but also works with smaller types.

Methods

impl<T> LocallyShuffledIds<T> where
T: One + Zero + Clone + Eq + Bounded + Default,
Wrapping<T>: Add<Wrapping<T>, Output = Wrapping<T>>,

fn new() -> LocallyShuffledIds<T>

Create a new LocallyShuffledIds struct.

fn start_at(start: T) -> LocallyShuffledIds<T>

Create a new LocallyShuffledIds struct at the starting value.